refactor: extract WikiPageBody from View.vue (#963 step 1)
Pure markdown-body renderer + click event handlers, used by
View.vue today and by the upcoming chat-inline preview
(WikiWriteResult, step 4). No behavior change — same `wiki-content`
class, same testids, same click semantics. Click handling becomes
emit-based: WikiPageBody emits taskCheckboxClick / wikiLinkClick /
workspaceLinkClick, View.vue wires those to its existing handlers.

Markdown→HTML pipeline lifts into renderWikiPageHtml() in helpers.ts
so the same transform is reusable from the chat preview without
copying the marked + image-rewrite + task-tagging chain.
